requirements.txt
alembic==0.9.9
altgraph==0.15
aniso8601==3.0.0
astroid==1.6.3
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap4==4.0.2
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
macholib==1.9
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pefile==2017.11.5
PyInstaller==3.3.1
pylint==1.8.4
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Continuum==1.3.4
SQLAlchemy-Utils==0.33.2
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
